Rename MarketAdvancedSearch classes to MarketFilters
